backend/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
cryptography
asgiref
pytest-asyncio
requests
markupsafe
google-genai